Use "long long" for PerfResult.total_bytes_downloaded to avoid overflow.
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet

+++ components/cronet/ios/test/cronet_performance_test.mm
@@ -32,7 +32,7 @@ struct PerfResult {
   NSTimeInterval total;
   NSTimeInterval mean;
   NSTimeInterval max;
-  int total_bytes_downloaded;
+  long long total_bytes_downloaded;
   int failed_requests;
   int total_requests;
 };
